requirements.txt
Flask==2.0.1
gunicorn==20.1.0
numpy==1.19.5
Pillow==7.0.0
requests==2.22.0
tensorflow-cpu==2.6.0
tensorflow-estimator==2.6.0